Shit, there are a lot of changes. Basically, matrix multiplication now works. This commit is going in before I add inner products and then do a quick n' dirty CG algorithm.
Also note that commons.math3.random.RandomGenerator does NOT work---fuck
them, I have no idea why. I spent hours doing it. I ended up using
something from MlLib, since there is already a good (modifiable)
generator in the RandomRDD class. Plus, anything written by Spark people
is generally better for Spark than the other crap on the internet.

Anyway, a helluva lot of changes, but the tests in:
~/tests/mult/mult-14.24.26-002
show that it works, and is fairly fast. Obviously I'm only comparing it
to the Pasalab marlin code, which is tested on a worse cluster. So
obviously it should be better. It seems like blocksizes should be chosen
based on thebest parallelism, and not on the actual size. After all, the
number crunching is short compared to the shuffle.

Peas out.